apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
alembic==1.13.3
pytest==8.3.4
pytz==2024.2
websockets==12.0
authlib==1.6.6
httpx==0.28.1
itsdangerous==2.2.0
